cleaned up some warnings
git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@5219 221aa14e-8319-0410-a670-987f0aec2ac5

+++ gr-atsc/src/lib/atsc_bit_timing_loop.cc
@@ -56,7 +56,7 @@ atsc_bit_timing_loop::forecast (int noutput_items, gr_vector_int &ninput_items_r
 {
   unsigned ninputs = ninput_items_required.size();
   for (unsigned i = 0; i < ninputs; i++)
-    ninput_items_required[i] = noutput_items * d_rx_clock_to_symbol_freq + 1500 - 1;
+    ninput_items_required[i] = static_cast<int>(noutput_items * d_rx_clock_to_symbol_freq) + 1500 - 1;
 
 }
